refactor: 将获取`ImageIndex`与`ImageToken`交给`BiliClient`管理
